election-research-agent/src/analysis/social_media_analyzer.py
from typing import Dict, List, Any
import logging
import tweepy
from langchain.chat_models import ChatOpenAI
from langchain.schema import HumanMessage, SystemMessage
import numpy as np
from datetime import datetime, timedelta

class SocialMediaAnalyzer:
    """Agent responsible for analyzing social media content related to elections."""
    
    def __init__(self):
        """Initialize the social media analyzer with necessary components."""
        self.logger = logging.getLogger(__name__)
        self.llm = ChatOpenAI(temperature=0)
        self._setup_twitter_client()
        
    def _setup_twitter_client(self):
        """Set up Twitter API client."""
        try:
            auth = tweepy.OAuthHandler(
                "YOUR_TWITTER_API_KEY",
                "YOUR_TWITTER_API_SECRET"
            )
            auth.set_access_token(
                "YOUR_ACCESS_TOKEN",
                "YOUR_ACCESS_TOKEN_SECRET"
            )
            self.twitter_client = tweepy.API(auth)
        except Exception as e:
            self.logger.error(f"Error setting up Twitter client: {str(e)}")
            raise
    
    def analyze(self, social_data: Dict[str, Any]) -> Dict[str, Any]:
        """
        Analyze social media content for election-related information.
        
        Args:
            social_data (Dict[str, Any]): Social media data including tweets and posts
            
        Returns:
            Dict[str, Any]: Analysis results including sentiment and trends
        """
        try:
            twitter_analysis = self._analyze_twitter(social_data.get('twitter', []))
            # Add other platforms as needed (Facebook, Instagram, etc.)
            
            return {
                'timestamp': datetime.now().isoformat(),
                'twitter_analysis': twitter_analysis,
                'overall_sentiment': self._calculate_overall_sentiment(twitter_analysis)
            }
        except Exception as e:
            self.logger.error(f"Error in social media analysis: {str(e)}")
            raise
    
    def _analyze_twitter(self, tweets: List[Dict[str, Any]]) -> Dict[str, Any]:
        """
        Analyze Twitter data using LLM.
        
        Args:
            tweets (List[Dict[str, Any]]): List of tweets and their metadata
            
        Returns:
            Dict[str, Any]: Analysis of Twitter content
        """
        try:
            # Prepare tweets for analysis
            tweet_texts = "\n\n".join([tweet['text'] for tweet in tweets])
            
            messages = [
                SystemMessage(content="You are an expert in social media analysis and political sentiment. Analyze these tweets for election-related patterns and sentiment."),
                HumanMessage(content=f"Tweets:\n{tweet_texts}")
            ]
            
            analysis = self.llm.invoke(messages)
            
            return {
                'sentiment_analysis': analysis.content,
                'tweet_count': len(tweets),
                'time_period': {
                    'start': min(tweet['created_at'] for tweet in tweets),
                    'end': max(tweet['created_at'] for tweet in tweets)
                }
            }
            
        except Exception as e:
            self.logger.error(f"Error analyzing Twitter data: {str(e)}")
            raise
    
    def _calculate_overall_sentiment(self, twitter_analysis: Dict[str, Any]) -> float:
        """
        Calculate overall sentiment score from various social media analyses.
        
        Args:
            twitter_analysis (Dict[str, Any]): Analysis results from Twitter
            
        Returns:
            float: Sentiment score between -1 and 1
        """
        try:
            # Here we would implement more sophisticated sentiment calculation
            # For now, return a placeholder neutral sentiment
            return 0.0
            
        except Exception as e:
            self.logger.error(f"Error calculating overall sentiment: {str(e)}")
            raise