election-research-agent/requirements.txt
# CPU-only configurations
--extra-index-url https://download.pytorch.org/whl/cpu
llama-cpp-python==0.2.11
beautifulsoup4>=4.12.2
requests>=2.31.0
tweepy>=4.14.0
pandas>=2.1.1
numpy>=1.24.3
python-dotenv>=1.0.0
selenium>=4.14.0
newspaper3k>=0.2.8
transformers>=4.34.0
torch>=2.1.0
scikit-learn>=1.3.2